Refactor world code into EngineCore, remove unused projects
