+++ src/Client/ClientCore/Rendering/Materials/RenderingMaterialManager.cs
@@ -28,9 +28,6 @@ using Sovereign.EngineCore.Main;
 using Sovereign.EngineCore.World.Materials;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Sovereign.ClientCore.Rendering.Materials
 {
