+++ src/Client/ClientCore/Rendering/Materials/RenderingMaterialValidator.cs
@@ -22,12 +22,10 @@
  */
 
 using Sovereign.ClientCore.Rendering.Sprites.TileSprites;
-using Sovereign.WorldLib.Materials;
-using System;
+using Sovereign.EngineCore.World.Materials;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace Sovereign.ClientCore.Rendering.Materials
 {
