+++ src/Client/ClientCore/Rendering/Scenes/Game/World/WorldLayerGrouper.cs
@@ -28,7 +28,6 @@ using Sovereign.EngineCore.Systems.Block.Components;
 using Sovereign.EngineCore.Systems.Movement.Components;
 using Sovereign.EngineCore.World.Materials;
 using Sovereign.EngineUtil.Collections;
-using Sovereign.WorldLib.Materials;
 using System.Collections.Generic;
 using System.Numerics;
