+++ src/Common/EngineCore/World/Materials/MaterialDefinitionsException.cs
@@ -23,37 +23,25 @@
 
 using System;
 using System.Collections.Generic;
+using System.Runtime.Serialization;
 using System.Text;
 
-namespace Sovereign.WorldLib.Materials
+namespace Sovereign.EngineCore.World.Materials;
+
+/// <summary>
+/// Exception type thrown when an error occurs with material definitions.
+/// </summary>
+public sealed class MaterialDefinitionsException : ApplicationException
 {
-
-    /// <summary>
-    /// Describes a subtype of a material.
-    /// </summary>
-    public class MaterialSubtype
+    public MaterialDefinitionsException()
     {
-
-        /// <summary>
-        /// The material modifier value. Unique within a material.
-        /// </summary>
-        public int MaterialModifier { get; set; }
-
-        /// <summary>
-        /// The ID of the tile sprite used for the top face.
-        /// </summary>
-        public int TopFaceTileSpriteId { get; set; }
-
-        /// <summary>
-        /// The ID of the tile sprite used for the top face if a face is obscured.
-        /// </summary>
-        public int ObscuredTopFaceTileSpriteId { get; set; }
-
-        /// <summary>
-        /// The ID of the tile sprite used for the side face.
-        /// </summary>
-        public int SideFaceTileSpriteId { get; set; }
-
     }
 
+    public MaterialDefinitionsException(string message) : base(message)
+    {
+    }
+
+    public MaterialDefinitionsException(string message, Exception innerException) : base(message, innerException)
+    {
+    }
 }
