--- src/Server/StandaloneWorldGen/Logging.xml
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<log4net>
-
-  <appender name="Console" type="log4net.Appender.ConsoleAppender">
-    <layout type="log4net.Layout.PatternLayout">
-      <conversionPattern value="%-4timestamp [%thread] %-5level %logger %ndc - %message%newline" />
-    </layout>
-  </appender>
-
-  <root>
-    <level value="DEBUG" />
-    <appender-ref ref="Console" />
-  </root>
-  
-</log4net>
