--- src/Server/StandaloneWorldGen/Properties/launchSettings.json
@@ -1,7 +1,0 @@
-{
-  "profiles": {
-    "StandaloneWorldGen": {
-      "commandName": "Project"
-    }
-  }
-}
