--- src/Server/StandaloneWorldGen/StandaloneWorldGen.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <StartupObject>Sovereign.StandaloneWorldGen.Main.StandaloneWorldGenMain</StartupObject>
-    <RootNamespace>Sovereign.StandaloneWorldGen</RootNamespace>
-    <AssemblyName>Sovereign.StandaloneWorldGen</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Logging.xml" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="Logging.xml">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\WorldGen\WorldGen.csproj" />
-  </ItemGroup>
-
-</Project>
