--- src/Server/WorldGen/Configuration/WorldGenConfigurationInstaller.cs
@@ -1,53 +1,0 @@
-﻿/*
- * Sovereign Engine
- * Copyright (c) 2018 opticfluorine
- *
- * Permission is hereby granted, free of charge, to any person obtaining a 
- * copy of this software and associated documentation files (the "Software"), 
- * to deal in the Software without restriction, including without limitation 
- * the rights to use, copy, modify, merge, publish, distribute, sublicense, 
- * and/or sell copies of the Software, and to permit persons to whom the 
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
- * DEALINGS IN THE SOFTWARE.
- */
-
-using Castle.MicroKernel.Registration;
-using Castle.MicroKernel.SubSystems.Configuration;
-using Castle.Windsor;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Sovereign.WorldGen.Configuration
-{
-
-    /// <summary>
-    /// IoC installer for WorldGen configuration helper classes.
-    /// </summary>
-    public class WorldGenConfigurationInstaller : IWindsorInstaller
-    {
-
-        public void Install(IWindsorContainer container, IConfigurationStore store)
-        {
-            /* WorldGenConfigurationLoader. */
-            container.Register(Component.For<WorldGenConfigurationLoader>()
-                .LifestyleTransient());
-
-            /* WorldGenConfigurationValidator. */
-            container.Register(Component.For<WorldGenConfigurationValidator>()
-                .LifestyleTransient());
-        }
-
-    }
-
-}
