--- src/Server/WorldGen/WorldGen.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>Sovereign.WorldGen</RootNamespace>
-    <AssemblyName>Sovereign.WorldGen</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Common\WorldLib\WorldLib.csproj" />
-  </ItemGroup>
-
-</Project>
